requirements.txt
google-api-python-client
oauth2client
google-cloud-resource-manager
ipcalc
requests